+++ src/components/App/App.scss
@@ -151,3 +151,4 @@
     }
 }
 }
+
