+++ src/components/Fetching/Fetching.scss
@@ -1,0 +1,13 @@
+.container-fetching{
+    height: 100vh;
+    width: 100%;
+    background-color:white;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.img-logo{
+    width: 20%;
+    height: auto;
+}
